testing/web-platform/tests/old-tests/submission/Infraware/Forms/contents/Forms/input_form.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
  <p>
    <h3>input_form</h3>
  </p>
  
  <hr>

  <div id="log"></div> 

  <form method="post"
      enctype="application/x-www-form-urlencoded"
      action=""
      name="input_form">
  <p><input type='hidden' id='input_text'></p>
  </form>

  <script>
    
    var input_text = document.getElementById("input_text");
    var form = input_text.form;

    if (typeof(form) == "object") {
      test(function() {
        assert_equals(form.name, 'input_form', "form attribute is not correct.");
      });
    } else {
      test(function() {
        assert_unreached("form attribute is not exist.");
      });
    }

  </script>

 </body>
</html>
